requirements.txt
Django==3.1.6
django-filebrowser==3.13.2
django-grappelli==2.14.3
Pillow==8.1.0
pytz==2020.1
gunicorn==20.0.4
